requirements.txt
ipython==8.5.0
Pillow==9.2.0
requests==2.28.1
